# simple-erp/erp-api/openapi.yaml
...
openapi: 3.0.3
info:
  title: REST API for study ERP system
  contact:
    name: Integrity Vision LLC
    url: https://integrity.com.ua
    email: vivanchenko@integrity.com.ua
  version: "1.0"
servers:
- url: /erp
paths:
  /api/rest/equipment:
    get:
      parameters:
      - name: type
        in: query
        schema:
          format: int64
          type: integer
      - name: withDecommissioned
        in: query
        schema:
          default: "false"
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipmentDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentDto'
  /api/rest/equipment/maintenance:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceDto'
  /api/rest/equipment/maintenance/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Return maintenance data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceDto'
        "404":
          description: Maintenance not exist
  /api/rest/equipment/maintenance/{id}/consumption-rate:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumptionRateDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceDto'
  /api/rest/equipment/maintenance/{id}/consumption-rate/{rateId}:
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: rateId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceDto'
  /api/rest/equipment/supply-rate:
    get:
      parameters:
      - name: type
        in: query
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupplyRateDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyRateDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyRateDto'
  /api/rest/equipment/supply-rate/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Return supply rate data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyRateDto'
        "404":
          description: Supply rate not exist
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyRateDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyRateDto'
  /api/rest/equipment/type:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipmentTypeDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentTypeDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentTypeDto'
  /api/rest/equipment/type/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Return equipment type data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentTypeDto'
        "404":
          description: Equipment type not exist
  /api/rest/equipment/type/{id}/consumption-rate:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumptionRateDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentTypeDto'
  /api/rest/equipment/type/{id}/consumption-rate/{rateId}:
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: rateId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentTypeDto'
  /api/rest/equipment/type/{id}/maintenance:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentTypeDto'
  /api/rest/equipment/type/{id}/maintenance/{maintenanceId}:
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: maintenanceId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentTypeDto'
  /api/rest/equipment/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Return equipment data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentDto'
        "404":
          description: Equipment not exist
  /api/rest/equipment/{id}/assign:
    patch:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentAssigningRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentDto'
  /api/rest/equipment/{id}/commission:
    patch:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCommissioningRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentDto'
  /api/rest/equipment/{id}/decommission:
    patch:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCommissioningRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentDto'
  /api/rest/equipment/{id}/retrieve:
    patch:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentDto'
...
components:
  schemas:
...
    ConsumptionRateDto:
      required:
      - title
      - material
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        material:
          allOf:
          - $ref: '#/components/schemas/MaterialDto'
          - nullable: false
        durationUnit:
          type: object
          xml:
            name: duration-unit
        durationAmount:
          format: int64
          type: integer
          xml:
            name: duration-amount
        amount:
          type: number
      xml:
        name: consumption-rate
...
    EquipmentAssigningRequest:
      required:
      - serialNumber
      - employeeId
      type: object
      properties:
        serialNumber:
          pattern: \S
          type: string
          xml:
            name: serial-number
          nullable: false
        employeeId:
          format: int64
          type: integer
          xml:
            name: employee-id
          nullable: false
      xml:
        name: equipment-assign-request
    EquipmentCommissioningRequest:
      required:
      - serialNumber
      - date
      type: object
      properties:
        serialNumber:
          pattern: \S
          type: string
          xml:
            name: serial-number
          nullable: false
        date:
          pattern: yyyy-MM-dd
          type: string
          nullable: false
      xml:
        name: equipment-commissioning-request
    EquipmentDto:
      required:
      - serialNumber
      - type
      - purchasePrice
      type: object
      properties:
        serialNumber:
          pattern: \S
          type: string
          xml:
            name: serial-number
          nullable: false
        type:
          format: int64
          type: integer
          nullable: false
        purchasePrice:
          type: number
          xml:
            name: purchase-price
          nullable: false
        balancePrice:
          type: number
          xml:
            name: balance-price
        commissioningDate:
          format: date
          type: string
          xml:
            name: commissioning-date
        decommissioningDate:
          format: date
          type: string
          xml:
            name: decommissioning-date
        user:
          format: int64
          type: integer
      xml:
        name: equipment
    EquipmentTypeDto:
      required:
      - title
      - lifeTimeInMonth
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        description:
          type: string
        lifeTimeInMonth:
          format: int32
          type: integer
          xml:
            name: life-time
          nullable: false
        maintenanceSet:
          uniqueItems: true
          type: array
          items:
            format: int64
            type: integer
            xml:
              name: maintenance-id
          xml:
            name: maintenance-set
            wrapped: true
        consumptionRates:
          uniqueItems: true
          type: array
          items:
            format: int64
            type: integer
            xml:
              name: rate-id
          xml:
            name: consumption-rates
            wrapped: true
      xml:
        name: equipment-type
    MaintenanceDto:
      required:
      - title
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        consumptionRates:
          uniqueItems: true
          type: array
          items:
            format: int64
            type: integer
            xml:
              name: rate-id
          xml:
            name: consumption-rates
            wrapped: true
      xml:
        name: maintenance
    MaterialDto:
      required:
      - title
      - measureUnit
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        measureUnit:
          allOf:
          - $ref: '#/components/schemas/MeasureUnit'
          - xml:
              name: measure-unit
            nullable: false
      xml:
        name: material
    MeasureUnit:
      enum:
      - METER
      - SQUARE_METER
      - KILOGRAMME
      - LITER
      - PCS
      - PKG
      type: string
...
    SupplyRateDto:
      required:
      - title
      - equipmentType
      - amount
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        equipmentType:
          format: int64
          type: integer
          xml:
            name: equipment-type
          nullable: false
        amount:
          format: int32
          type: integer
          nullable: false
      xml:
        name: supply-rate
...